refactor(2d-render-pipeline): fix import error
