+++ packages/core/src/2d/text/CharRenderInfo.ts
@@ -2,7 +2,7 @@ import { Vector4 } from "@galacean/engine-math";
 import { Texture2D } from "../../texture";
 import { Chunk } from "../../RenderPipeline/Chunk";
 import { Engine } from "../../Engine";
-import { IPoolElement } from "../../utils/ReturnableObjectPool";
+import { IPoolElement } from "../../utils/ObjectPool";
 
 /**
  * @internal
