+++ packages/core/src/RenderPipeline/RenderElement.ts
@@ -1,5 +1,5 @@
 import { ShaderPass } from "../shader/ShaderPass";
-import { IPoolElement } from "../utils/ReturnableObjectPool";
+import { IPoolElement } from "../utils/ObjectPool";
 import { RenderData } from "./RenderData";
 
 export class RenderElement implements IPoolElement {
