+++ packages/core/src/graphic/SubMesh.ts
@@ -1,4 +1,4 @@
-import { IPoolElement } from "../utils/ReturnableObjectPool";
+import { IPoolElement } from "../utils/ObjectPool";
 import { MeshTopology } from "./enums/MeshTopology";
 
 /**
